django_rest_framework/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers

